refactor: dao 디렉토리명을 entity로 변경
