+++ src/main/java/com/samsam/travel/travelcommerce/entity/Orders.java
@@ -1,4 +1,4 @@
-package com.samsam.travel.travelcommerce.dao;
+package com.samsam.travel.travelcommerce.entity;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Min;
